added the main.h file

+++ 0x03-debugging/main.h
@@ -3,6 +3,6 @@
 
 #include <stdio.h>
 void positive_or_negative(int i);
-long int largest_number(long int a, long int b, long int c);
+int largest_number(int a, int b, int c);
 
 #endif /* MAIN_H */
